+++ enginesdl.h
@@ -32,7 +32,7 @@ class EngineSDL : public Engine
     	bool isSupported() { return true; } ///< Since SDL is always supported, this function always returns true.
 		void doResize(int& h, int& w);
 
-		void drawLightmap(vertex* lightmap, int type, int width, int height, int scale);
+		void drawLightmap(vertex* lightmap, int type, int width, int height, float scale);
     	void drawRectangle(float x, float y, float width, float height, oRGBA color);
     	void drawRectangleLines(float x, float y, float width, float height, oRGBA color, float thickness = 1.f);
